refactor(CO-3211): update api type imports to use specific paths for better clarity
