+++ src/tests/app.test.tsx
@@ -23,24 +23,7 @@ import * as registerShellIntegrations from 'app-utils/register-shell-integration
 import * as useSearchRegisterer from 'app-utils/use-search-registerer';
 import { BACKUP_SEARCH_ROUTE } from 'constants/index';
 import { useBackupSearchStore } from 'store/backup-search/store';
-import { DeletedMessageFromAPI } from 'types';
-
-// Mocking the worker. In commons jest-setup the worker is already mocked, but is improperly defined with wrong types and
-// is causing a call to "onMessage", which tries to alter the folders store and overrides the folders, breaking the test.
-// It also causes warning/errors due the fact it tries to set an "undefined" in the folders.
-// I think we should consider removing that mock or redefine it or make it configurable
-// vi.mock('@zextras/carbonio-ui-commons', async () => ({
-// 	...(await vi.importActual('@zextras/carbonio-ui-commons')),
-// 	folderWorker: {
-// 		postMessage: vi.fn()
-// 	},
-// 	tagsWorker: {
-// 		postMessage: vi.fn()
-// 	}
-// }));
-// vi.mock('@zextras/carbonio-shell-ui', async () => ({
-// 	...(await vi.importActual('@zextras/carbonio-shell-ui'))
-// }));
+import { DeletedMessageFromAPI } from 'types/api';
 
 function aDeletedMessage(): DeletedMessageFromAPI {
 	return {
